// web/src/api/generated/schemas/moduleSummary.zod.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Fonzytooter API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from 'zod'

export const ModuleSummary = zod.strictObject({
  id: zod.string(),
  order: zod.int(),
  title: zod.string(),
})

export type ModuleSummary = zod.input<typeof ModuleSummary>
export type ModuleSummaryOutput = zod.output<typeof ModuleSummary>